up: update some app run logic
